Split out tests as well
