+++ test/builtin/associations/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests code under module test.builtin.associations"""
